Animations, transitions & user experience

+++ TodoList-MVVM/Views/ListView.swift
@@ -11,23 +11,33 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
 
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            } else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
+                    .onDelete(perform: listViewModel.deleteItem)
+                    .onMove(perform: listViewModel.moveItem)
+                }
+                .listStyle(PlainListStyle())
             }
-            .onDelete(perform: listViewModel.deleteItem)
-            .onMove(perform: listViewModel.moveItem)
         }
-        .listStyle(PlainListStyle())
-        .navigationTitle("Todo List 📋")
+
+        .navigationTitle("📋 Todo List")
         .toolbar {
             ToolbarItem(placement: .navigationBarLeading) {
-                EditButton()
+                if !listViewModel.items.isEmpty {
+                    EditButton()
+                }
             }
             ToolbarItem(placement: .navigationBarTrailing) {
                 NavigationLink("Add") {
